♻️ refactor: TypeORM 파일을 도메인 모듈 안으로 이동
