+++ services/api/app/src/api/public/campuses/campuses.service.ts
@@ -1,7 +1,7 @@
 import { Cache, CACHE_MANAGER } from '@nestjs/cache-manager';
 import { Inject, Injectable, Logger } from '@nestjs/common';
 import { CampusListResult } from 'src/api/public/campuses/dtos/results/campus-list-result.dto';
-import { CampusesRepository } from 'src/type-orm/entities/campuses/campuses.repository';
+import { CampusesRepository } from 'src/api/public/campuses/campuses.repository';
 import { CacheKey } from 'src/api/common/decorators/cache-key.decorator';
 
 @Injectable()
