+++ services/api/app/src/api/public/departments/departments.service.ts
@@ -1,7 +1,7 @@
 import { Cache, CACHE_MANAGER } from '@nestjs/cache-manager';
 import { Inject, Injectable, Logger } from '@nestjs/common';
 import { DepartmentListResult } from 'src/api/public/departments/dtos/results/department-list-result.dto';
-import { DepartmentsRepository } from 'src/type-orm/entities/departments/departments.repository';
+import { DepartmentsRepository } from 'src/api/public/departments/departments.repository';
 import { CacheKey } from 'src/api/common/decorators/cache-key.decorator';
 
 @Injectable()
