+++ services/api/app/src/api/public/notices/entities/notice.entity.ts
