+++ services/api/app/src/api/public/notices/notices.module.ts
@@ -1,14 +1,24 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { CommonMessageFactory } from 'src/api/public/common/common-message.factory';
+import { NoticeCategory } from 'src/api/public/notices/entities/notice-category.entity';
+import { Notice } from 'src/api/public/notices/entities/notice.entity';
+import { NoticeCategoriesRepository } from 'src/api/public/notices/notice-categories.repository';
 import { NoticeMessageFactory } from 'src/api/public/notices/notice-message.factory';
-import { NoticesRepositoryModule } from 'src/type-orm/entities/notices/notices-repository.module';
+import { NoticesRepository } from 'src/api/public/notices/notices.repository';
 import { NoticesController } from './notices.controller';
 import { NoticesService } from './notices.service';
 
 @Module({
-  imports: [NoticesRepositoryModule],
+  imports: [TypeOrmModule.forFeature([Notice, NoticeCategory])],
   controllers: [NoticesController],
-  providers: [NoticesService, NoticeMessageFactory, CommonMessageFactory],
+  providers: [
+    NoticesService,
+    NoticesRepository,
+    NoticeCategoriesRepository,
+    NoticeMessageFactory,
+    CommonMessageFactory,
+  ],
   exports: [NoticesService],
 })
 export class NoticesModule {}
