+++ services/api/app/src/api/public/notices/notices.service.ts
@@ -3,11 +3,11 @@ import {
   NoticeCategoryResult,
   NoticeListResult,
 } from 'src/api/public/notices/dtos/results/notice-list-result.dto';
-import { NoticeCategoriesRepository } from 'src/type-orm/entities/notices/notice-categories.repository';
-import { NoticeCategory } from 'src/type-orm/entities/notices/notice-category.entity';
-import { Notice } from 'src/type-orm/entities/notices/notice.entity';
-import { NoticesRepository } from 'src/type-orm/entities/notices/notices.repository';
-import { User } from 'src/type-orm/entities/users/users.entity';
+import { NoticeCategoriesRepository } from 'src/api/public/notices/notice-categories.repository';
+import { NoticeCategory } from 'src/api/public/notices/entities/notice-category.entity';
+import { Notice } from 'src/api/public/notices/entities/notice.entity';
+import { NoticesRepository } from 'src/api/public/notices/notices.repository';
+import { User } from 'src/api/public/users/entities/users.entity';
 
 @Injectable()
 export class NoticesService {
