+++ services/api/app/src/api/public/schedules/entities/academic-calendar.entity.ts
