+++ services/api/app/src/api/public/schedules/schedules.module.ts
@@ -1,14 +1,21 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { CommonMessageFactory } from 'src/api/public/common/common-message.factory';
+import { AcademicCalendarsRepository } from 'src/api/public/schedules/academic-calendars.repository';
+import { AcademicCalendar } from 'src/api/public/schedules/entities/academic-calendar.entity';
 import { ScheduleMessageFactory } from 'src/api/public/schedules/schedule-message.factory';
-import { AcademicCalendarsRepositoryModule } from 'src/type-orm/entities/academic-calendars/academic-calendars-repository.module';
 import { SchedulesController } from 'src/api/public/schedules/schedules.controller';
 import { SchedulesService } from './schedules.service';
 
 @Module({
-  imports: [AcademicCalendarsRepositoryModule],
+  imports: [TypeOrmModule.forFeature([AcademicCalendar])],
   controllers: [SchedulesController],
-  providers: [SchedulesService, ScheduleMessageFactory, CommonMessageFactory],
+  providers: [
+    SchedulesService,
+    AcademicCalendarsRepository,
+    ScheduleMessageFactory,
+    CommonMessageFactory,
+  ],
   exports: [SchedulesService],
 })
 export class SchedulesModule {}
