+++ services/api/app/src/api/public/shuttles/entities/shuttle-timetable.entity.ts
