+++ services/api/app/src/api/public/shuttles/shuttle-timetable.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { ShuttleTimetable } from './shuttle-timetable.entity';
+import { ShuttleTimetable } from './entities/shuttle-timetable.entity';
 
 @Injectable()
 export class ShuttleTimetableRepository {
