+++ services/api/app/src/api/public/shuttles/shuttles.module.ts
@@ -1,8 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DatabaseModule } from 'src/type-orm/database.module';
-import { ShuttleTimetable } from 'src/type-orm/entities/shuttle-timetables/shuttle-timetable.entity';
-import { ShuttleTimetableRepository } from 'src/type-orm/entities/shuttle-timetables/shuttle-timetable.repository';
+import { ShuttleTimetable } from 'src/api/public/shuttles/entities/shuttle-timetable.entity';
+import { ShuttleTimetableRepository } from 'src/api/public/shuttles/shuttle-timetable.repository';
 import { ShuttleMessageFactory } from './shuttle-message.factory';
 import { ShuttleTimetableCalculator } from './shuttle-timetable.calculator';
 import { ShuttlesController } from './shuttles.controller';
@@ -10,7 +9,7 @@ import { ShuttlesNativeController } from './shuttles-native.controller';
 import { ShuttlesService } from './shuttles.service';
 
 @Module({
-  imports: [DatabaseModule, TypeOrmModule.forFeature([ShuttleTimetable])],
+  imports: [TypeOrmModule.forFeature([ShuttleTimetable])],
   controllers: [ShuttlesController, ShuttlesNativeController],
   providers: [
     ShuttlesService,
