+++ services/api/app/src/api/public/users/users.module.ts
@@ -5,9 +5,8 @@ import { CampusesModule } from 'src/api/public/campuses/campuses.module';
 import { CollegesModule } from 'src/api/public/colleges/colleges.module';
 import { DepartmentsModule } from 'src/api/public/departments/departments.module';
 import { CurrentUserInterceptor } from 'src/api/public/users/interceptors/current-user.interceptor';
-import { DatabaseModule } from '../../../type-orm/database.module';
-import { User } from '../../../type-orm/entities/users/users.entity';
-import { UsersRepository } from '../../../type-orm/entities/users/users.repository';
+import { User } from 'src/api/public/users/entities/users.entity';
+import { UsersRepository } from 'src/api/public/users/users.repository';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { UserMessageFactory } from 'src/api/public/users/user-message.factory';
@@ -17,13 +16,7 @@ import { CollegeMessageFactory } from 'src/api/public/colleges/college-message.f
 import { DepartmentMessageFactory } from 'src/api/public/departments/department-message.factory';
 
 @Module({
-  imports: [
-    DatabaseModule,
-    TypeOrmModule.forFeature([User]),
-    CampusesModule,
-    CollegesModule,
-    DepartmentsModule,
-  ],
+  imports: [TypeOrmModule.forFeature([User]), CampusesModule, CollegesModule, DepartmentsModule],
   controllers: [UsersController],
   providers: [
     UsersService,
