+++ services/api/app/src/api/public/users/users.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { UpsertDepartmentResult } from 'src/api/public/users/dtos/results/upsert-department-result.dto';
 import { UserProfileResult } from 'src/api/public/users/dtos/results/user-profile-result.dto';
-import { User } from '../../../type-orm/entities/users/users.entity';
-import { UsersRepository } from '../../../type-orm/entities/users/users.repository';
+import { User } from 'src/api/public/users/entities/users.entity';
+import { UsersRepository } from 'src/api/public/users/users.repository';
 
 @Injectable()
 export class UsersService {
